Move Diagram to uml package
