+++ src/main/java/nl/talsmasoftware/umldoclet/uml/Diagram.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.talsmasoftware.umldoclet.diagrams;
+package nl.talsmasoftware.umldoclet.uml;
 
 import net.sourceforge.plantuml.FileFormat;
 import net.sourceforge.plantuml.FileFormatOption;
